backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic (Data validation)
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment & Config
python-dotenv==1.0.0

# Supabase
supabase==2.3.4
postgrest==0.13.2

# Anthropic (Claude API)
anthropic==0.8.1

# HTTP Client (for Runway API)
httpx==0.24.1  
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Rate Limiting
slowapi==0.1.9

# CORS
python-multipart==0.0.6

# Monitoring (optional, production)
sentry-sdk[fastapi]==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality (development)
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Database (async support)
asyncpg==0.29.0